requirements.txt
cohere==4.32
langchain==0.0.319
moviepy==1.0.3
openai==0.28.1
pandas==2.1.1
pinecone-client==2.2.4
python-dotenv==1.0.0
st-click-detector==0.1.3
streamlit==1.28.0
